requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
# awsclis
flake8
python-dotenv>=0.5.1
torch>=1.8
torchvision
matplotlib